Added Ability to Browse Directories
For directories, I added a "Browse" link which will send the user
to a new "Browse" view. Both the route and view are copies of the
respective "Index" route and view. These should be refactored but
some slight differences between using the IPFS API when browsing vs
the initial index via ipfs-cluster API make it easier to copy for
now.

Would also be nice to maintain path for breadcrumbs.