Fixed Incorrect Display of UnixFS Files as Directories
Previously, I was checking if an object had links to determine if
it was a directory; however, unixfs objects can be chunked into
multiple blocks which was causing false positives. Now upon finding
links, I get the data for the object and then unmarshal it into a
unixfs object to determine its type.